Adding auto.cc plus minor changes

+++ c++/src/sizeof.cc
@@ -4,11 +4,11 @@ int main()
 {
     const char *str = "test";
 
-    printf("sizeof(\"test\"): %llu\n", sizeof("test"));
+    printf("sizeof(\"test\"): %llu\n", (unsigned long long) sizeof("test"));
 
-    printf("sizeof(str): %llu\n", sizeof(str));
+    printf("sizeof(str): %llu\n", (unsigned long long) sizeof(str));
 
-    printf("sizeof(*str): %llu\n", sizeof(*str));
+    printf("sizeof(*str): %llu\n", (unsigned long long) sizeof(*str));
 
     return 0;
 }
